% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_piracy.R
\docType{data}
\name{dt_piracy}
\alias{dt_piracy}
\title{Piracy dataset}
\format{A data table with 1271400 rows and 3 variables:
\describe{
  \item{sa}{service account}
  \item{imdbID}{item identifier used by imdb}
  \item{n_events}{number of times this sa was caught downloading this imdbID}
}}
\usage{
dt_piracy
}
\description{
A dataset containing downloads of 189048 users across 19370 items.
}
\keyword{datasets}

