% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendations.R
\name{randomRec}
\alias{randomRec}
\title{Random recommendations}
\usage{
randomRec(data, n, modelsize = 1000, testsize = 500, partsize = 10000,
  eval.alg = F, keepModel = F, oldmodel = NULL)
}
\arguments{
\item{data}{A binaryRatingMatrix.}

\item{n}{The number of recommendations we want for each user.}

\item{modelsize}{The maximum size for the model.}

\item{testsize}{The maximum size for the test set.}

\item{partsize}{The size for prediction partitions.}

\item{eval.alg}{If TRUE the function will return the evaluation confusion matrix}

\item{keepModel}{If TRUE the function will return the computed model}

\item{oldmodel}{A previously computed model to be used for predicitons}
}
\value{
A three element list:
 [[1]] contains the recommendations;
 [[2]] contains the model if \code{keepModel} is TRUE;
 [[3]] contains a confusion matrix if \code{eval.alg} is TRUE.
}
\description{
Random recommendations
}
\examples{
randomRec(data=ratings.sample,n=5)
randomRec(data=ratings.sample,n=5,modelsize=20000,testsize=1000,partsize=10000,eval.alg=T,keepModel=T)

}

