% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendations.R
\name{recommend}
\alias{recommend}
\title{Recommendations base function}
\usage{
recommend(data, alg, n = 1, parameter = NULL, progress = F,
  keepModel = F, oldmodel = NULL, modelsize = 40000, testsize = 1000,
  partsize = 10000, reduction = F, eval.alg = F, newdata = NULL)
}
\arguments{
\item{data}{A binaryRatingMatrix.}

\item{alg}{The algorithm we will use to get recommendations.}

\item{n}{The number of recommendations we want for each user.}

\item{parameter}{Parameters specific to the algorithm we will use.}

\item{progress}{if TRUE the time usage for each step will be printed.}

\item{keepModel}{If TRUE the function will return the computed model.}

\item{oldmodel}{A previously computed model to be used for predicitons.}

\item{modelsize}{The maximum size for the model.}

\item{testsize}{The maximum size for the test set.}

\item{partsize}{The size for prediction partitions.}

\item{reduction}{If TRUE we will use dimensionality reduction on the train set.}

\item{eval.alg}{If TRUE the function will return the evaluation confusion matrix}
}
\value{
A three element list:
 [[1]] contains the recommendations;
 [[2]] contains the model if \code{keepModel} is TRUE;
 [[3]] contains a confusion matrix if \code{eval.alg} is TRUE.
}
\description{
Recommendations base function
}
\examples{
recommend(data=ratings.sample,alg="POPULAR",modelsize=20000,testsize=1000,partsize=10000,eval.alg=T,n=5,keepModel=T)

}

