% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommenderyang.R
\docType{package}
\name{recommenderyang}
\alias{recommenderyang}
\alias{recommenderyang-package}
\title{recommnderyang: A package to obtain recommendations.}
\description{
The recommenderyang package provides several popular algorithms to
obtain recommendations
}
\section{Algorithms}{


Random
Popular
IBCF
UBCF
UBCF (with dimnsionality reduction)
SVD
}

